+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Trait to decrease the cost casting force/tech powers.
+- Support for Supreme Accuracy / Aptitude / Durability traits.
 
 ### Changes